Split iterator-test into iterator-test and iterator-seek-test
